Working ResumableUpload checksumming and unit tests
