+++ tests/unit/requests/test_download.py
@@ -62,7 +62,7 @@ class TestDownload(object):
         chunk2 = b"second chunk, or chunk 1, which is better? "
         chunk3 = b"ordinals and numerals and stuff."
         header_value = u"crc32c=qmNCyg==,md5=fPAJHnnoi/+NadyNxT2c2w=="
-        headers = {download_mod._HASH_HEADER: header_value}
+        headers = {_root_helpers._HASH_HEADER: header_value}
         response = _mock_response(chunks=[chunk1, chunk2, chunk3], headers=headers)
 
         ret_val = download._write_to_stream(response)
@@ -87,7 +87,7 @@ class TestDownload(object):
         chunk3 = b"ordinals and numerals and stuff."
         bad_checksum = u"d3JvbmcgbiBtYWRlIHVwIQ=="
         header_value = u"crc32c={bad},md5={bad}".format(bad=bad_checksum)
-        headers = {download_mod._HASH_HEADER: header_value}
+        headers = {_root_helpers._HASH_HEADER: header_value}
         response = _mock_response(chunks=[chunk1, chunk2, chunk3], headers=headers)
 
         with pytest.raises(common.DataCorruption) as exc_info:
@@ -127,7 +127,7 @@ class TestDownload(object):
         chunk3 = b"ordinals and numerals and stuff."
         bad_checksum = u"d3JvbmcgbiBtYWRlIHVwIQ=="
         header_value = u"crc32c={bad},md5={bad}".format(bad=bad_checksum)
-        headers = {download_mod._HASH_HEADER: header_value}
+        headers = {_root_helpers._HASH_HEADER: header_value}
         response = _mock_response(chunks=[chunk1, chunk2, chunk3], headers=headers)
 
         with pytest.raises(ValueError) as exc_info:
@@ -211,7 +211,7 @@ class TestDownload(object):
         stream = io.BytesIO()
         chunks = (b"up down ", b"charlie ", b"brown")
         header_value = u"crc32c=UNIQxg==,md5=JvS1wjMvfbCXgEGeaJJLDQ=="
-        headers = {download_mod._HASH_HEADER: header_value}
+        headers = {_root_helpers._HASH_HEADER: header_value}
         transport = self._consume_helper(
             stream=stream, chunks=chunks, response_headers=headers, checksum=checksum
         )
@@ -234,7 +234,7 @@ class TestDownload(object):
         chunks = (b"zero zero", b"niner tango")
         bad_checksum = u"anVzdCBub3QgdGhpcyAxLA=="
         header_value = u"crc32c={bad},md5={bad}".format(bad=bad_checksum)
-        headers = {download_mod._HASH_HEADER: header_value}
+        headers = {_root_helpers._HASH_HEADER: header_value}
         transport = mock.Mock(spec=["request"])
         transport.request.return_value = _mock_response(chunks=chunks, headers=headers)
 
@@ -309,7 +309,7 @@ class TestRawDownload(object):
         chunk2 = b"second chunk, or chunk 1, which is better? "
         chunk3 = b"ordinals and numerals and stuff."
         header_value = u"crc32c=qmNCyg==,md5=fPAJHnnoi/+NadyNxT2c2w=="
-        headers = {download_mod._HASH_HEADER: header_value}
+        headers = {_root_helpers._HASH_HEADER: header_value}
         response = _mock_raw_response(chunks=[chunk1, chunk2, chunk3], headers=headers)
 
         ret_val = download._write_to_stream(response)
@@ -336,7 +336,7 @@ class TestRawDownload(object):
         chunk3 = b"ordinals and numerals and stuff."
         bad_checksum = u"d3JvbmcgbiBtYWRlIHVwIQ=="
         header_value = u"crc32c={bad},md5={bad}".format(bad=bad_checksum)
-        headers = {download_mod._HASH_HEADER: header_value}
+        headers = {_root_helpers._HASH_HEADER: header_value}
         response = _mock_raw_response(chunks=[chunk1, chunk2, chunk3], headers=headers)
 
         with pytest.raises(common.DataCorruption) as exc_info:
@@ -376,7 +376,7 @@ class TestRawDownload(object):
         chunk3 = b"ordinals and numerals and stuff."
         bad_checksum = u"d3JvbmcgbiBtYWRlIHVwIQ=="
         header_value = u"crc32c={bad},md5={bad}".format(bad=bad_checksum)
-        headers = {download_mod._HASH_HEADER: header_value}
+        headers = {_root_helpers._HASH_HEADER: header_value}
         response = _mock_response(chunks=[chunk1, chunk2, chunk3], headers=headers)
 
         with pytest.raises(ValueError) as exc_info:
@@ -460,7 +460,7 @@ class TestRawDownload(object):
         stream = io.BytesIO()
         chunks = (b"up down ", b"charlie ", b"brown")
         header_value = u"crc32c=UNIQxg==,md5=JvS1wjMvfbCXgEGeaJJLDQ=="
-        headers = {download_mod._HASH_HEADER: header_value}
+        headers = {_root_helpers._HASH_HEADER: header_value}
         transport = self._consume_helper(
             stream=stream, chunks=chunks, response_headers=headers, checksum=checksum
         )
@@ -485,7 +485,7 @@ class TestRawDownload(object):
         chunks = (b"zero zero", b"niner tango")
         bad_checksum = u"anVzdCBub3QgdGhpcyAxLA=="
         header_value = u"crc32c={bad},md5={bad}".format(bad=bad_checksum)
-        headers = {download_mod._HASH_HEADER: header_value}
+        headers = {_root_helpers._HASH_HEADER: header_value}
         transport = mock.Mock(spec=["request"])
         transport.request.return_value = _mock_raw_response(
             chunks=chunks, headers=headers
@@ -735,129 +735,6 @@ class TestRawChunkedDownload(object):
         assert download.total_bytes == total_bytes
 
 
-class Test__get_expected_checksum(object):
-    @pytest.mark.parametrize("checksum", [u"md5", u"crc32c"])
-    @mock.patch("google.resumable_media.requests.download._LOGGER")
-    def test__w_header_present(self, _LOGGER, checksum):
-        checksums = {"md5": u"b2twdXNodGhpc2J1dHRvbg==", "crc32c": u"3q2+7w=="}
-        header_value = u"crc32c={},md5={}".format(checksums["crc32c"], checksums["md5"])
-        headers = {download_mod._HASH_HEADER: header_value}
-        response = _mock_response(headers=headers)
-
-        def _get_headers(response):
-            return response.headers
-
-        expected_checksum, checksum_obj = download_mod._get_expected_checksum(
-            response, _get_headers, EXAMPLE_URL, checksum_type=checksum
-        )
-        assert expected_checksum == checksums[checksum]
-
-        checksum_types = {
-            "md5": type(hashlib.md5()),
-            "crc32c": type(_root_helpers._get_crc32c_object()),
-        }
-        assert isinstance(checksum_obj, checksum_types[checksum])
-
-        _LOGGER.info.assert_not_called()
-
-    @pytest.mark.parametrize("checksum", [u"md5", u"crc32c"])
-    @mock.patch("google.resumable_media.requests.download._LOGGER")
-    def test__w_header_missing(self, _LOGGER, checksum):
-        headers = {}
-        response = _mock_response(headers=headers)
-
-        def _get_headers(response):
-            return response.headers
-
-        expected_checksum, checksum_obj = download_mod._get_expected_checksum(
-            response, _get_headers, EXAMPLE_URL, checksum_type=checksum
-        )
-        assert expected_checksum is None
-        assert isinstance(checksum_obj, download_mod._DoNothingHash)
-        expected_msg = download_mod._MISSING_CHECKSUM.format(
-            EXAMPLE_URL, checksum_type=checksum.upper()
-        )
-        _LOGGER.info.assert_called_once_with(expected_msg)
-
-
-class Test__parse_checksum_header(object):
-
-    CRC32C_CHECKSUM = u"3q2+7w=="
-    MD5_CHECKSUM = u"c2l4dGVlbmJ5dGVzbG9uZw=="
-
-    def test_empty_value(self):
-        header_value = None
-        response = None
-        md5_header = download_mod._parse_checksum_header(
-            header_value, response, checksum_label="md5"
-        )
-        assert md5_header is None
-        crc32c_header = download_mod._parse_checksum_header(
-            header_value, response, checksum_label="crc32c"
-        )
-        assert crc32c_header is None
-
-    def test_crc32c_only(self):
-        header_value = u"crc32c={}".format(self.CRC32C_CHECKSUM)
-        response = None
-        md5_header = download_mod._parse_checksum_header(
-            header_value, response, checksum_label="md5"
-        )
-        assert md5_header is None
-        crc32c_header = download_mod._parse_checksum_header(
-            header_value, response, checksum_label="crc32c"
-        )
-        assert crc32c_header == self.CRC32C_CHECKSUM
-
-    def test_md5_only(self):
-        header_value = u"md5={}".format(self.MD5_CHECKSUM)
-        response = None
-        md5_header = download_mod._parse_checksum_header(
-            header_value, response, checksum_label="md5"
-        )
-        assert md5_header == self.MD5_CHECKSUM
-        crc32c_header = download_mod._parse_checksum_header(
-            header_value, response, checksum_label="crc32c"
-        )
-        assert crc32c_header is None
-
-    def test_both_crc32c_and_md5(self):
-        header_value = u"crc32c={},md5={}".format(
-            self.CRC32C_CHECKSUM, self.MD5_CHECKSUM
-        )
-        response = None
-        md5_header = download_mod._parse_checksum_header(
-            header_value, response, checksum_label="md5"
-        )
-        assert md5_header == self.MD5_CHECKSUM
-        crc32c_header = download_mod._parse_checksum_header(
-            header_value, response, checksum_label="crc32c"
-        )
-        assert crc32c_header == self.CRC32C_CHECKSUM
-
-    def test_md5_multiple_matches(self):
-        another_checksum = u"eW91IGRpZCBXQVQgbm93Pw=="
-        header_value = u"md5={},md5={}".format(self.MD5_CHECKSUM, another_checksum)
-        response = mock.sentinel.response
-
-        with pytest.raises(common.InvalidResponse) as exc_info:
-            download_mod._parse_checksum_header(
-                header_value, response, checksum_label="md5"
-            )
-
-        error = exc_info.value
-        assert error.response is response
-        assert len(error.args) == 3
-        assert error.args[1] == header_value
-        assert error.args[2] == [self.MD5_CHECKSUM, another_checksum]
-
-
-def test__DoNothingHash():
-    do_nothing_hash = download_mod._DoNothingHash()
-    return_value = do_nothing_hash.update(b"some data")
-    assert return_value is None
-
-
 class Test__add_decoder(object):
     def test_non_gzipped(self):
         response_raw = mock.Mock(headers={}, spec=["headers"])
@@ -871,7 +748,7 @@ class Test__add_decoder(object):
         md5_hash = download_mod._add_decoder(response_raw, mock.sentinel.md5_hash)
 
         assert md5_hash is not mock.sentinel.md5_hash
-        assert isinstance(md5_hash, download_mod._DoNothingHash)
+        assert isinstance(md5_hash, _root_helpers._DoNothingHash)
         assert isinstance(response_raw._decoder, download_mod._GzipDecoder)
         assert response_raw._decoder._checksum is mock.sentinel.md5_hash
